Adding a config variable for specifying cloud provider of postgres database
Some cloud providers' implementations of their PostgreSQL offering have some nuances and require workarounds for postgres-operator to function properly.
