--- pkg/postgres/queries.go
@@ -1,20 +1,0 @@
-package postgres
-
-const (
-	CREATE_DB            = `CREATE DATABASE "%s"`
-	CREATE_SCHEMA        = `CREATE SCHEMA IF NOT EXISTS "%s" AUTHORIZATION "%s"`
-	ALTER_DB_OWNER       = `ALTER DATABASE "%s" OWNER TO "%s"`
-	CREATE_GROUP_ROLE    = `CREATE ROLE "%s"`
-	CREATE_USER_ROLE     = `CREATE ROLE "%s" WITH LOGIN PASSWORD '%s'`
-	GRANT_ROLE           = `GRANT "%s" TO "%s"`
-	ALTER_USER_SET_ROLE  = `ALTER USER "%s" SET ROLE "%s"`
-	REVOKE_ROLE          = `REVOKE "%s" FROM "%s"`
-	UPDATE_PASSWORD      = `ALTER ROLE "%s" WITH PASSWORD '%s'`
-	DROP_ROLE            = `DROP ROLE "%s"`
-	DROP_DATABASE        = `DROP DATABASE "%s"`
-	DROP_OWNED_BY        = `DROP OWNED BY "%s"`
-	REASIGN_OBJECTS      = `REASSIGN OWNED BY "%s" TO "%s"`
-	GRANT_USAGE_SCHEMA   = `GRANT USAGE ON SCHEMA "%s" TO "%s"`
-	GRANT_ALL_TABLES     = `GRANT %s ON ALL TABLES IN SCHEMA "%s" TO "%s"`
-	DEFAULT_PRIVS_SCHEMA = `ALTER DEFAULT PRIVILEGES FOR ROLE "%s" IN SCHEMA "%s" GRANT %s ON TABLES TO "%s"`
-)
